Refactor CLI helpers into focused modules
